requirements.txt
nltk==3.5
python-dotenv==0.14.0
requests == 2.28.2
scikit-learn==0.23.2
vk-api==11.9.4
